+++ src/create_objects/setup_scene/setup_objects.c
@@ -6,7 +6,7 @@
 /*   By: rnijhuis <rnijhuis@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/05/25 19:46:15 by rnijhuis      #+#    #+#                 */
-/*   Updated: 2022/06/27 23:00:01 by rubennijhui   ########   odam.nl         */
+/*   Updated: 2022/06/30 20:32:01 by rubennijhui   ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -55,8 +55,8 @@ void	make_light(t_light *light, char *settings)
 
 void	set_lights(t_scene *scene, t_line *file_content)
 {
-	uint32_t	current_light;
-	uint32_t	i;
+	size_t	current_light;
+	size_t	i;
 
 	i = 0;
 	scene->lights = ft_calloc(scene->amount_lights, sizeof(t_light));
@@ -76,7 +76,7 @@ void	set_lights(t_scene *scene, t_line *file_content)
 
 t_line	*find_obj_in_file(char *definition, t_line *file_content)
 {
-	uint32_t	current_line;
+	size_t	current_line;
 
 	current_line = 0;
 	while (file_content[current_line].line != NULL)
